+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## Addeed
+* A type alias `Option[T]` for `Some[T]`.
+
 
 ## 0.0.1
 ## Added